move shared files into parent diretory
