+++ src/rapi/rc_api_editor.c
@@ -1,7 +1,7 @@
 #include "rc_api_editor.h"
 #include "rc_api_common.h"
 
-#include "../rcheevos/rc_compat.h"
+#include "../rc_compat.h"
 #include "../rhash/md5.h"
 
 #include <stdlib.h>
