+++ src/rc_client_internal.h
