+++ src/rc_compat.c
