+++ src/rc_compat.h
