+++ test/rapi/test_rc_api_info.c
@@ -2,7 +2,7 @@
 
 #include "../src/rapi/rc_api_common.h"
 #include "../test_framework.h"
-#include "rc_compat.h"
+#include "../rc_compat.h"
 
 #define DOREQUEST_URL "https://retroachievements.org/dorequest.php"
