+++ test/rapi/test_rc_api_user.c
@@ -2,8 +2,8 @@
 
 #include "../src/rapi/rc_api_common.h"
 #include "../test_framework.h"
-#include "rc_compat.h"
-#include "rc_version.h"
+#include "../rc_compat.h"
+#include "../rc_version.h"
 
 #define DOREQUEST_URL "https://retroachievements.org/dorequest.php"
