+++ test/rcheevos-test.vcxproj
@@ -132,7 +132,6 @@
     <ClCompile Include="..\src\rapi\rc_api_runtime.c" />
     <ClCompile Include="..\src\rapi\rc_api_user.c" />
     <ClCompile Include="..\src\rcheevos\alloc.c" />
-    <ClCompile Include="..\src\rcheevos\compat.c" />
     <ClCompile Include="..\src\rcheevos\condition.c" />
     <ClCompile Include="..\src\rcheevos\condset.c" />
     <ClCompile Include="..\src\rcheevos\consoleinfo.c" />
@@ -140,14 +139,15 @@
     <ClCompile Include="..\src\rcheevos\lboard.c" />
     <ClCompile Include="..\src\rcheevos\memref.c" />
     <ClCompile Include="..\src\rcheevos\operand.c" />
-    <ClCompile Include="..\src\rcheevos\rc_libretro.c" />
-    <ClCompile Include="..\src\rcheevos\rc_client.c" />
     <ClCompile Include="..\src\rcheevos\rc_validate.c" />
     <ClCompile Include="..\src\rcheevos\richpresence.c" />
     <ClCompile Include="..\src\rcheevos\runtime.c" />
     <ClCompile Include="..\src\rcheevos\runtime_progress.c" />
     <ClCompile Include="..\src\rcheevos\trigger.c" />
     <ClCompile Include="..\src\rcheevos\value.c" />
+    <ClCompile Include="..\src\rc_client.c" />
+    <ClCompile Include="..\src\rc_compat.c" />
+    <ClCompile Include="..\src\rc_libretro.c" />
     <ClCompile Include="..\src\rhash\cdreader.c" />
     <ClCompile Include="..\src\rhash\hash.c" />
     <ClCompile Include="..\src\rhash\md5.c" />
@@ -197,11 +197,9 @@
     <ClCompile Include="rcheevos\test_lboard.c" />
     <ClCompile Include="rcheevos\test_memref.c" />
     <ClCompile Include="rcheevos\test_operand.c" />
-    <ClCompile Include="rcheevos\test_rc_libretro.c" />
     <ClCompile Include="rcheevos\test_rc_validate.c" />
     <ClCompile Include="rcheevos\test_richpresence.c" />
     <ClCompile Include="rcheevos\test_runtime.c" />
-    <ClCompile Include="rcheevos\test_rc_client.c" />
     <ClCompile Include="rcheevos\test_runtime_progress.c" />
     <ClCompile Include="rcheevos\test_timing.c" />
     <ClCompile Include="rcheevos\test_trigger.c" />
@@ -212,6 +210,8 @@
     <ClCompile Include="rhash\test_hash.c" />
     <ClCompile Include="rurl\test_url.c" />
     <ClCompile Include="test.c" />
+    <ClCompile Include="test_rc_client.c" />
+    <ClCompile Include="test_rc_libretro.c" />
     <ClInclude Include="..\include\rcheevos.h" />
     <ClInclude Include="..\include\rc_api_editor.h" />
     <ClInclude Include="..\include\rc_api_info.h" />
@@ -224,14 +224,14 @@
     <ClInclude Include="..\include\rc_client.h" />
     <ClInclude Include="..\include\rc_runtime_types.h" />
     <ClInclude Include="..\include\rc_url.h" />
-    <ClInclude Include="..\src\rcheevos\rc_compat.h" />
     <ClInclude Include="..\src\rcheevos\rc_internal.h" />
     <ClInclude Include="..\include\rc_error.h" />
     <ClInclude Include="..\src\rapi\rc_api_common.h" />
-    <ClInclude Include="..\src\rcheevos\rc_libretro.h" />
-    <ClInclude Include="..\src\rcheevos\rc_client_internal.h" />
     <ClInclude Include="..\src\rcheevos\rc_validate.h" />
-    <ClInclude Include="..\src\rcheevos\rc_version.h" />
+    <ClInclude Include="..\src\rc_client_internal.h" />
+    <ClInclude Include="..\src\rc_compat.h" />
+    <ClInclude Include="..\src\rc_libretro.h" />
+    <ClInclude Include="..\src\rc_version.h" />
     <ClInclude Include="..\src\rhash\md5.h" />
     <ClInclude Include="libretro.h" />
     <ClInclude Include="rcheevos\mock_memory.h" />
