+++ test/rcheevos-test.vcxproj.filters
@@ -210,9 +210,6 @@
     <ClCompile Include="rcheevos\test_richpresence.c">
       <Filter>tests\rcheevos</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\rcheevos\compat.c">
-      <Filter>src\rcheevos</Filter>
-    </ClCompile>
     <ClCompile Include="rcheevos\test_runtime.c">
       <Filter>tests\rcheevos</Filter>
     </ClCompile>
@@ -261,12 +258,6 @@
     <ClCompile Include="rapi\test_rc_api_common.c">
       <Filter>tests\rapi</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\rcheevos\rc_libretro.c">
-      <Filter>src\rcheevos</Filter>
-    </ClCompile>
-    <ClCompile Include="rcheevos\test_rc_libretro.c">
-      <Filter>tests\rcheevos</Filter>
-    </ClCompile>
     <ClCompile Include="..\src\rapi\rc_api_info.c">
       <Filter>src\rapi</Filter>
     </ClCompile>
@@ -288,11 +279,20 @@
     <ClCompile Include="rcheevos\test_timing.c">
       <Filter>tests\rcheevos</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\rcheevos\rc_client.c">
-      <Filter>src\rcheevos</Filter>
+    <ClCompile Include="..\src\rc_client.c">
+      <Filter>src</Filter>
     </ClCompile>
-    <ClCompile Include="rcheevos\test_rc_client.c">
-      <Filter>tests\rcheevos</Filter>
+    <ClCompile Include="..\src\rc_compat.c">
+      <Filter>src</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\rc_libretro.c">
+      <Filter>src</Filter>
+    </ClCompile>
+    <ClCompile Include="test_rc_client.c">
+      <Filter>tests</Filter>
+    </ClCompile>
+    <ClCompile Include="test_rc_libretro.c">
+      <Filter>tests</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -326,9 +326,6 @@
     <ClInclude Include="..\src\rcheevos\rc_internal.h">
       <Filter>src\rcheevos</Filter>
     </ClInclude>
-    <ClInclude Include="..\src\rcheevos\rc_compat.h">
-      <Filter>src\rcheevos</Filter>
-    </ClInclude>
     <ClInclude Include="..\include\rc_error.h">
       <Filter>src\rcheevos</Filter>
     </ClInclude>
@@ -343,9 +340,6 @@
     </ClInclude>
     <ClInclude Include="..\include\rc_api_runtime.h">
       <Filter>src\rapi</Filter>
-    </ClInclude>
-    <ClInclude Include="..\src\rcheevos\rc_libretro.h">
-      <Filter>src\rcheevos</Filter>
     </ClInclude>
     <ClInclude Include="libretro.h">
       <Filter>tests\rcheevos</Filter>
@@ -365,14 +359,20 @@
     <ClInclude Include="..\src\rcheevos\rc_validate.h">
       <Filter>src\rcheevos</Filter>
     </ClInclude>
-    <ClInclude Include="..\src\rcheevos\rc_version.h">
-      <Filter>src\rcheevos</Filter>
-    </ClInclude>
     <ClInclude Include="..\include\rc_client.h">
       <Filter>src\rcheevos</Filter>
     </ClInclude>
-    <ClInclude Include="..\src\rcheevos\rc_client_internal.h">
-      <Filter>src\rcheevos</Filter>
+    <ClInclude Include="..\src\rc_version.h">
+      <Filter>src</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\rc_client_internal.h">
+      <Filter>src</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\rc_compat.h">
+      <Filter>src</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\rc_libretro.h">
+      <Filter>src</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
