+++ test/rcheevos/test_rc_validate.c
@@ -1,8 +1,8 @@
 #include "rc_validate.h"
 
-#include "rc_compat.h"
 #include "rc_consoles.h"
 
+#include "../rc_compat.h"
 #include "../test_framework.h"
 
 #include <stdlib.h>
