+++ test/rcheevos/test_richpresence.c
@@ -3,7 +3,7 @@
 #include "../test_framework.h"
 #include "mock_memory.h"
 
-#include "../src/rcheevos/rc_compat.h"
+#include "../src/rc_compat.h"
 
 static void _assert_parse_richpresence(rc_richpresence_t** richpresence, void* buffer, const char* script) {
   int size;
