+++ test/rhash/mock_filereader.c
@@ -1,6 +1,6 @@
 #include "rc_hash.h"
 
-#include "../rcheevos/rc_compat.h"
+#include "../rc_compat.h"
 
 #include <stdlib.h>
 #include <string.h>
