+++ test/test_rc_client.c
@@ -3,12 +3,13 @@
 #include "rc_consoles.h"
 #include "rc_hash.h"
 #include "rc_internal.h"
-#include "rc_client_internal.h"
-#include "rc_version.h"
 #include "rc_api_runtime.h"
 
-#include "../rhash/data.h"
-#include "../test_framework.h"
+#include "../src/rc_client_internal.h"
+#include "../src/rc_version.h"
+
+#include "rhash/data.h"
+#include "test_framework.h"
 
 #if defined(_WIN32)
 #include <Windows.h>
